+++ core/app_administration.go
@@ -404,14 +404,16 @@ func (app *application) admRemoveAccountsFromGroup(appID string, orgID string, g
 
 	//ensure that the accounts have the group
 	updateAccounts := make([]model.Account, 0)
+	hasPermissions := make([]bool, 0)
 	for _, account := range accounts {
 		if account.GetGroup(groupID) != nil {
 			updateAccounts = append(updateAccounts, account)
+			hasPermissions = append(hasPermissions, len(account.Permissions) > 0 || len(account.Roles) > 0 || len(account.Groups) > 1)
 		}
 	}
 
 	//remove the accounts from the group
-	err = app.storage.RemoveAccountsGroup(group.ID, updateAccounts)
+	err = app.storage.RemoveAccountsGroup(group.ID, updateAccounts, hasPermissions)
 	if err != nil {
 		return errors.Wrapf("error removing accounts from a group - %s", err, groupID)
 	}
@@ -544,9 +546,10 @@ func (app *application) admGetApplicationPermissions(appID string, orgID string,
 	return permissions, nil
 }
 
-func (app *application) admGetAccounts(appID string, orgID string, accountID *string, authTypeIdentifier *string) ([]model.Account, error) {
+func (app *application) admGetAccounts(limit int, offset int, appID string, orgID string, accountID *string, firstName *string, lastName *string, authType *string,
+	authTypeIdentifier *string, hasPermissions *bool, permissions []string, roleIDs []string, groupIDs []string) ([]model.Account, error) {
 	//find the accounts
-	accounts, err := app.storage.FindAccounts(appID, orgID, accountID, authTypeIdentifier)
+	accounts, err := app.storage.FindAccounts(limit, offset, appID, orgID, accountID, firstName, lastName, authType, authTypeIdentifier, hasPermissions, permissions, roleIDs, groupIDs)
 	if err != nil {
 		return nil, errors.WrapErrorAction(logutils.ActionFind, model.TypeAccount, nil, err)
 	}
@@ -670,12 +673,15 @@ func (app *application) admRevokeAccountPermissions(appID string, orgID string, 
 	}
 
 	//verify that the account has the permissions which are supposed to be revoked
+	numRevoked := 0
 	for _, current := range permissionNames {
 		hasP := account.GetPermissionNamed(current)
 		if hasP == nil {
 			l.Infof("trying to revoke %s for %s but the account does not have it", current, accountID)
 			return errors.Newf("%s cannot be revoked from %s", current, accountID)
 		}
+
+		numRevoked++
 	}
 
 	//find permissions
@@ -695,10 +701,11 @@ func (app *application) admRevokeAccountPermissions(appID string, orgID string, 
 		}
 	}
 
+	hasPermissions := len(account.Permissions) > numRevoked || len(account.Roles) > 0 || len(account.Groups) > 0
 	//delete permissions from an account AND delete all sessions for the account
 	transaction := func(context storage.TransactionContext) error {
 		//delete permissions from an account
-		err = app.storage.DeleteAccountPermissions(context, accountID, permissions)
+		err = app.storage.DeleteAccountPermissions(context, accountID, hasPermissions, permissions)
 		if err != nil {
 			return errors.Wrap("error deleting account permissions", err)
 		}
@@ -774,12 +781,15 @@ func (app *application) admRevokeAccountRoles(appID string, orgID string, accoun
 	}
 
 	//verify that the account has the roles which are supposed to be revoked
+	numRevoked := 0
 	for _, roleID := range roleIDs {
 		hasR := account.GetRole(roleID)
 		if hasR == nil {
 			l.Infof("trying to revoke role %s for %s but the account does not have it", roleID, accountID)
 			return errors.Newf("%s cannot be revoked from %s", roleID, accountID)
 		}
+
+		numRevoked++
 	}
 
 	//find roles
@@ -799,10 +809,11 @@ func (app *application) admRevokeAccountRoles(appID string, orgID string, accoun
 		}
 	}
 
+	hasPermissions := len(account.Permissions) > 0 || len(account.Roles) > numRevoked || len(account.Groups) > 0
 	//delete roles from an account AND delete all sessions for the account
 	transaction := func(context storage.TransactionContext) error {
 		//delete roles from an account
-		err = app.storage.DeleteAccountRoles(context, accountID, roleIDs)
+		err = app.storage.DeleteAccountRoles(context, accountID, hasPermissions, roleIDs)
 		if err != nil {
 			return errors.Wrap("error deleting account roles", err)
 		}