+++ core/model/application.go
@@ -122,10 +122,10 @@ func (c AppOrgRole) CheckAssigners(assignerPermissions []string) error {
 	for _, permission := range c.Permissions {
 		err := permission.CheckAssigners(assignerPermissions)
 		if err != nil {
-			errors.Wrapf("error checking role permission assigners", err)
+			return errors.Wrapf("error checking role permission assigners", err)
 		}
 	}
-	//it satisies all permissions
+	//it satisfies all permissions
 	return nil
 }