+++ driven/storage/model_user.go
@@ -25,9 +25,10 @@ type account struct {
 
 	AppOrgID string `bson:"app_org_id,omitempty"`
 
-	Permissions []model.Permission `bson:"permissions,omitempty"`
-	Roles       []accountRole      `bson:"roles,omitempty"`
-	Groups      []accountGroup     `bson:"groups,omitempty"`
+	HasPermissions bool               `bson:"has_permissions"`
+	Permissions    []model.Permission `bson:"permissions,omitempty"`
+	Roles          []accountRole      `bson:"roles,omitempty"`
+	Groups         []accountGroup     `bson:"groups,omitempty"`
 
 	AuthTypes []accountAuthType `bson:"auth_types,omitempty"`