+++ driver/web/conversions_user.go
@@ -17,6 +17,7 @@ package web
 import (
 	"core-building-block/core/model"
 	Def "core-building-block/driver/web/docs/gen"
+	"core-building-block/utils"
 )
 
 //Account
@@ -33,7 +34,7 @@ func accountToDef(item model.Account) *Def.SharedResAccount {
 	groups := accountGroupsToDef(item.GetActiveGroups())
 	//account auth types
 	authTypes := accountAuthTypesToDef(item.AuthTypes)
-	return &Def.SharedResAccount{Id: item.ID, Permissions: &permissions, Roles: &roles, Groups: &groups,
+	return &Def.SharedResAccount{Id: item.ID, HasPermissions: &item.HasPermissions, Permissions: &permissions, Roles: &roles, Groups: &groups,
 		AuthTypes: &authTypes, Profile: profile, Preferences: preferences}
 }
 
@@ -59,9 +60,9 @@ func partialAccountToDef(item model.Account, params map[string]interface{}) *Def
 	}
 	//dates
 	var dateUpdated *string
-	dateCreated := item.DateCreated.Format("2006-01-02T15:04:05.000Z")
+	dateCreated := utils.FormatTime(&item.DateCreated)
 	if item.DateUpdated != nil {
-		formatted := item.DateUpdated.Format("2006-01-02T15:04:05.000Z")
+		formatted := utils.FormatTime(item.DateUpdated)
 		dateUpdated = &formatted
 	}
 
@@ -71,8 +72,16 @@ func partialAccountToDef(item model.Account, params map[string]interface{}) *Def
 		paramsData = &params
 	}
 	return &Def.PartialAccount{Id: item.ID, AppId: item.AppOrg.Application.ID, OrgId: item.AppOrg.Organization.ID,
-		FirstName: item.Profile.FirstName, LastName: item.Profile.LastName, Permissions: permissions, Roles: roles, Groups: groups,
+		FirstName: item.Profile.FirstName, LastName: item.Profile.LastName, HasPermissions: item.HasPermissions, Permissions: permissions, Roles: roles, Groups: groups,
 		AuthTypes: authTypes, DateCreated: dateCreated, DateUpdated: dateUpdated, Params: paramsData}
+}
+
+func partialAccountsToDef(items []model.Account) []Def.PartialAccount {
+	result := make([]Def.PartialAccount, len(items))
+	for i, item := range items {
+		result[i] = *partialAccountToDef(item, nil)
+	}
+	return result
 }
 
 //AccountAuthType