+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Prevent admins from using service account management endpoints [#500](https://github.com/rokwire/core-building-block/issues/500)
 
 ### Changed
+- Get all admin level accounts [#486](https://github.com/rokwire/core-building-block/issues/486)
 - Update SECURITY.md [#504](https://github.com/rokwire/core-building-block/issues/504)
 
 ## [1.23.0] - 2022-04-26